requirements.txt
django-filter
djangorestframework
asgiref==3.2.7
Django==3.0.7
django-crispy-forms==1.9.1
django-material==1.6.7
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
